refac: extract pipe controllers
